\name{SUMS-package}
\alias{SUMS-package}
\alias{SUMS}
\docType{package}
\title{
\packageTitle{SUMS}
}
\description{
\packageDescription{SUMS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SUMS}
\packageIndices{SUMS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SUMS}

Maintainer: \packageMaintainer{SUMS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
